+++ crates/native/src/api/mod.rs
@@ -14,6 +14,7 @@ pub mod rtcp_feedback;
 pub mod rtp_capabilities;
 pub mod rtp_codec_capability;
 pub mod rtp_header_extension_capability;
+pub mod rtp_transceiver_init;
 
 use std::{
     sync::{
@@ -66,6 +67,7 @@ pub use self::{
     },
     rtp_codec_capability::{RtpCodecCapability, set_codec_preferences},
     rtp_header_extension_capability::RtpHeaderExtensionCapability,
+    rtp_transceiver_init::{RtpTransceiverDirection, RtpTransceiverInit},
 };
 // Re-exporting since it is used in the generated code.
 pub use crate::{
@@ -651,100 +653,6 @@ impl From<sys::PeerConnectionState> for PeerConnectionState {
             _ => unreachable!(),
         }
     }
-}
-
-/// [RTCRtpTransceiverDirection][1] representation.
-///
-/// [1]: https://w3.org/TR/webrtc#dom-rtcrtptransceiverdirection
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RtpTransceiverDirection {
-    /// The [RTCRtpTransceiver]'s [RTCRtpSender] will offer to send RTP, and
-    /// will send RTP if the remote peer accepts. The [RTCRtpTransceiver]'s
-    /// [RTCRtpReceiver] will offer to receive RTP, and will receive RTP if the
-    /// remote peer accepts.
-    ///
-    /// [RTCRtpReceiver]: https://w3.org/TR/webrtc#dom-rtcrtpreceiver
-    /// [RTCRtpSender]: https://w3.org/TR/webrtc#dom-rtcrtpsender
-    /// [RTCRtpTransceiver]: https://w3.org/TR/webrtc#dom-rtcrtptransceiver
-    SendRecv,
-
-    /// The [RTCRtpTransceiver]'s [RTCRtpSender] will offer to send RTP, and
-    /// will send RTP if the remote peer accepts. The [RTCRtpTransceiver]'s
-    /// [RTCRtpReceiver] will not offer to receive RTP, and will not receive
-    /// RTP.
-    ///
-    /// [RTCRtpReceiver]: https://w3.org/TR/webrtc#dom-rtcrtpreceiver
-    /// [RTCRtpSender]: https://w3.org/TR/webrtc#dom-rtcrtpsender
-    /// [RTCRtpTransceiver]: https://w3.org/TR/webrtc#dom-rtcrtptransceiver
-    SendOnly,
-
-    /// The [RTCRtpTransceiver]'s [RTCRtpSender] will not offer to send RTP,
-    /// and will not send RTP. The [RTCRtpTransceiver]'s [RTCRtpReceiver] will
-    /// offer to receive RTP, and will receive RTP if the remote peer accepts.
-    ///
-    /// [RTCRtpReceiver]: https://w3.org/TR/webrtc#dom-rtcrtpreceiver
-    /// [RTCRtpSender]: https://w3.org/TR/webrtc#dom-rtcrtpsender
-    /// [RTCRtpTransceiver]: https://w3.org/TR/webrtc#dom-rtcrtptransceiver
-    RecvOnly,
-
-    /// The [RTCRtpTransceiver]'s [RTCRtpSender] will not offer to send RTP,
-    /// and will not send RTP. The [RTCRtpTransceiver]'s [RTCRtpReceiver] will
-    /// not offer to receive RTP, and will not receive RTP.
-    ///
-    /// [RTCRtpReceiver]: https://w3.org/TR/webrtc#dom-rtcrtpreceiver
-    /// [RTCRtpSender]: https://w3.org/TR/webrtc#dom-rtcrtpsender
-    /// [RTCRtpTransceiver]: https://w3.org/TR/webrtc#dom-rtcrtptransceiver
-    Inactive,
-
-    /// The [RTCRtpTransceiver] will neither send nor receive RTP. It will
-    /// generate a zero port in the offer. In answers, its [RTCRtpSender] will
-    /// not offer to send RTP, and its [RTCRtpReceiver] will not offer to
-    /// receive RTP. This is a terminal state.
-    ///
-    /// [RTCRtpReceiver]: https://w3.org/TR/webrtc#dom-rtcrtpreceiver
-    /// [RTCRtpSender]: https://w3.org/TR/webrtc#dom-rtcrtpsender
-    /// [RTCRtpTransceiver]: https://w3.org/TR/webrtc#dom-rtcrtptransceiver
-    Stopped,
-}
-
-impl From<sys::RtpTransceiverDirection> for RtpTransceiverDirection {
-    fn from(state: sys::RtpTransceiverDirection) -> Self {
-        match state {
-            sys::RtpTransceiverDirection::kSendRecv => Self::SendRecv,
-            sys::RtpTransceiverDirection::kSendOnly => Self::SendOnly,
-            sys::RtpTransceiverDirection::kRecvOnly => Self::RecvOnly,
-            sys::RtpTransceiverDirection::kInactive => Self::Inactive,
-            sys::RtpTransceiverDirection::kStopped => Self::Stopped,
-            _ => unreachable!(),
-        }
-    }
-}
-
-impl From<RtpTransceiverDirection> for sys::RtpTransceiverDirection {
-    fn from(state: RtpTransceiverDirection) -> Self {
-        match state {
-            RtpTransceiverDirection::SendRecv => Self::kSendRecv,
-            RtpTransceiverDirection::SendOnly => Self::kSendOnly,
-            RtpTransceiverDirection::RecvOnly => Self::kRecvOnly,
-            RtpTransceiverDirection::Inactive => Self::kInactive,
-            RtpTransceiverDirection::Stopped => Self::kStopped,
-        }
-    }
-}
-
-/// Representation of an [RTCRtpTransceiverInit][0].
-///
-/// [0]: https://w3.org/TR/webrtc#dom-rtcrtptransceiverinit
-pub struct RtpTransceiverInit {
-    /// Direction of the [RTCRtpTransceiver].
-    ///
-    /// [RTCRtpTransceiver]: https://w3.org/TR/webrtc#dom-rtcrtptransceiver
-    pub direction: RtpTransceiverDirection,
-
-    /// Sequence containing parameters for sending [RTP] encodings of media.
-    ///
-    /// [RTP]: https://en.wikipedia.org/wiki/Real-time_Transport_Protocol
-    pub send_encodings: Vec<RtcRtpEncodingParameters>,
 }
 
 /// Representation of a permanent pair of an [RTCRtpSender] and an
