+++ crates/native/src/api/rtp_transceiver_init.rs
@@ -1,0 +1,101 @@
+//! See [RTCRtpTransceiverInit][0].
+//!
+//! [0]: https://w3.org/TR/webrtc#dom-rtcrtptransceiverinit
+
+use libwebrtc_sys as sys;
+
+use crate::api::RtcRtpEncodingParameters;
+
+/// [RTCRtpTransceiverDirection][1] representation.
+///
+/// [1]: https://w3.org/TR/webrtc#dom-rtcrtptransceiverdirection
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub enum RtpTransceiverDirection {
+    /// The [RTCRtpTransceiver]'s [RTCRtpSender] will offer to send RTP, and
+    /// will send RTP if the remote peer accepts. The [RTCRtpTransceiver]'s
+    /// [RTCRtpReceiver] will offer to receive RTP, and will receive RTP if the
+    /// remote peer accepts.
+    ///
+    /// [RTCRtpReceiver]: https://w3.org/TR/webrtc#dom-rtcrtpreceiver
+    /// [RTCRtpSender]: https://w3.org/TR/webrtc#dom-rtcrtpsender
+    /// [RTCRtpTransceiver]: https://w3.org/TR/webrtc#dom-rtcrtptransceiver
+    SendRecv,
+
+    /// The [RTCRtpTransceiver]'s [RTCRtpSender] will offer to send RTP, and
+    /// will send RTP if the remote peer accepts. The [RTCRtpTransceiver]'s
+    /// [RTCRtpReceiver] will not offer to receive RTP, and will not receive
+    /// RTP.
+    ///
+    /// [RTCRtpReceiver]: https://w3.org/TR/webrtc#dom-rtcrtpreceiver
+    /// [RTCRtpSender]: https://w3.org/TR/webrtc#dom-rtcrtpsender
+    /// [RTCRtpTransceiver]: https://w3.org/TR/webrtc#dom-rtcrtptransceiver
+    SendOnly,
+
+    /// The [RTCRtpTransceiver]'s [RTCRtpSender] will not offer to send RTP,
+    /// and will not send RTP. The [RTCRtpTransceiver]'s [RTCRtpReceiver] will
+    /// offer to receive RTP, and will receive RTP if the remote peer accepts.
+    ///
+    /// [RTCRtpReceiver]: https://w3.org/TR/webrtc#dom-rtcrtpreceiver
+    /// [RTCRtpSender]: https://w3.org/TR/webrtc#dom-rtcrtpsender
+    /// [RTCRtpTransceiver]: https://w3.org/TR/webrtc#dom-rtcrtptransceiver
+    RecvOnly,
+
+    /// The [RTCRtpTransceiver]'s [RTCRtpSender] will not offer to send RTP,
+    /// and will not send RTP. The [RTCRtpTransceiver]'s [RTCRtpReceiver] will
+    /// not offer to receive RTP, and will not receive RTP.
+    ///
+    /// [RTCRtpReceiver]: https://w3.org/TR/webrtc#dom-rtcrtpreceiver
+    /// [RTCRtpSender]: https://w3.org/TR/webrtc#dom-rtcrtpsender
+    /// [RTCRtpTransceiver]: https://w3.org/TR/webrtc#dom-rtcrtptransceiver
+    Inactive,
+
+    /// The [RTCRtpTransceiver] will neither send nor receive RTP. It will
+    /// generate a zero port in the offer. In answers, its [RTCRtpSender] will
+    /// not offer to send RTP, and its [RTCRtpReceiver] will not offer to
+    /// receive RTP. This is a terminal state.
+    ///
+    /// [RTCRtpReceiver]: https://w3.org/TR/webrtc#dom-rtcrtpreceiver
+    /// [RTCRtpSender]: https://w3.org/TR/webrtc#dom-rtcrtpsender
+    /// [RTCRtpTransceiver]: https://w3.org/TR/webrtc#dom-rtcrtptransceiver
+    Stopped,
+}
+
+impl From<sys::RtpTransceiverDirection> for RtpTransceiverDirection {
+    fn from(state: sys::RtpTransceiverDirection) -> Self {
+        match state {
+            sys::RtpTransceiverDirection::kSendRecv => Self::SendRecv,
+            sys::RtpTransceiverDirection::kSendOnly => Self::SendOnly,
+            sys::RtpTransceiverDirection::kRecvOnly => Self::RecvOnly,
+            sys::RtpTransceiverDirection::kInactive => Self::Inactive,
+            sys::RtpTransceiverDirection::kStopped => Self::Stopped,
+            _ => unreachable!(),
+        }
+    }
+}
+
+impl From<RtpTransceiverDirection> for sys::RtpTransceiverDirection {
+    fn from(state: RtpTransceiverDirection) -> Self {
+        match state {
+            RtpTransceiverDirection::SendRecv => Self::kSendRecv,
+            RtpTransceiverDirection::SendOnly => Self::kSendOnly,
+            RtpTransceiverDirection::RecvOnly => Self::kRecvOnly,
+            RtpTransceiverDirection::Inactive => Self::kInactive,
+            RtpTransceiverDirection::Stopped => Self::kStopped,
+        }
+    }
+}
+
+/// Representation of an [RTCRtpTransceiverInit][0].
+///
+/// [0]: https://w3.org/TR/webrtc#dom-rtcrtptransceiverinit
+pub struct RtpTransceiverInit {
+    /// Direction of the [RTCRtpTransceiver].
+    ///
+    /// [RTCRtpTransceiver]: https://w3.org/TR/webrtc#dom-rtcrtptransceiver
+    pub direction: RtpTransceiverDirection,
+
+    /// Sequence containing parameters for sending [RTP] encodings of media.
+    ///
+    /// [RTP]: https://en.wikipedia.org/wiki/Real-time_Transport_Protocol
+    pub send_encodings: Vec<RtcRtpEncodingParameters>,
+}
