+++ lib/src/api/bridge/frb.dart
@@ -23,3 +23,4 @@ export 'api/rtcp_feedback.dart';
 export 'api/rtp_capabilities.dart';
 export 'api/rtp_codec_capability.dart';
 export 'api/rtp_header_extension_capability.dart';
+export 'api/rtp_transceiver_init.dart';
